[TASK] Make fluid_styled_content a framework extension
This is important to not have overhead in autoload generation
in non composer mode. The classes for this extension are registered in
composer.json of typo3/cms so they can be skipped when generating
additional class maps in non composer mode. They are skipped once
we change the package type to typo3-cms-framework

Resolves: #70085
Releases: master
Change-Id: I97747b84073d358a37c066e2110b7b2d5431d587
Reviewed-on: http://review.typo3.org/43528
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>