[BUGFIX] fsc/csc: Fix typolink parameters in parseFunc
This is a simple typo in the lib.parseFunc of both fluid_styled_content
and css_styled_content. The ATagParams attribute for the typolink
function needs to be written with a capital A in order to work,
otherwise all additional parameters on <a> tags will be dropped in the
rendered output.

Resolves: #79896
Releases: master
Change-Id: I54d5c18fe6934c03b97ae197aba56055394916e5
Reviewed-on: https://review.typo3.org/51755
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Sascha Egerer <sascha@sascha-egerer.de>
Tested-by: Sascha Egerer <sascha@sascha-egerer.de>