[!!!][TASK] Remove deprecated parsing of <link> tag
The original <link> tag was replaced in TYPO3 v8 by using
proper <a> tags with href="t3://" syntax, which is handled
by the new LinkService introduced in v8. All database RTE fields
were migrated to <a> tags, but the link tag code was kept for
the time being, which is now removed.

The result is that <link> tags are now never replaced
when rendering the Frontend now.

Resolves: #87755
Releases: master
Change-Id: I10afba763736a6b91cbca513f7496dee86da8902
Reviewed-on: https://review.typo3.org/c/59775
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Joerg Kummer <typo3@enobe.de>
Reviewed-by: Josef Glatz <josef.glatz@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>