[TASK] Deprecate parseFunc.sword functionality
The parseFunc.sword functionality wraps HTML (usually a span tag)
around RTE-based content if:
* GET parameter ?sword[]=search-word is set
* no_cache=1 is set via GET parameter or TypoScript

This feature only works in non-cached environments due to the actual
problem that the search term should never be cached.

This happens if a result page in indexed search is clicked, which
then receives a no_cache=1 argument. This feature is highly discouraged
due to the possible server load that is created, and thus removed.

The existing solution is far from optimal and should rather be handled in
third-party extensions (such as EXT:solr does it via a HighlightViewHelper).

Resolves: #95318
Releases: master
Change-Id: If74247bad14a87c55d092322f0a609c3663f2e67
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71176
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>