[TASK] Avoid setUpBackendUserFromFixture() in functionals
The testing-framework method setUpBackendUserFromFixture()
is kinda flawed since it does two things at once: It
creates db records and creates global state. This is
unfortunate in various cases, especially when  db
snapshoting is used. Additionally, it relies on a
class property which defaults to a fixture file
delivered by testing-framework in XML format. We
want to get rid of the .xml fixtures, and we don't
want that the testing-framework delivers default
fixtures.

The patch adapts all usages of the property and
of method setUpBackendUserFromFixture() by splitting
it into two calls to first insert a local extension
fixture .csv file and to then call the backend user
init helper method.

This finishes the removal of .xml fixture uses
in the core.

$ composer req --dev typo3/testing-framework:dev-main

Change-Id: I32a5c43a6c8176a9e9182d08e08b42b5ff279634
Resolves: #97928
Related: #96856
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75168
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>